Greek TIN Number Validation
The Greek Tax Identification Number (TIN) consists of 9 digits. The last digit is a check digit. It is calculated as follows:
We remove the check digit, so that we are left with an 8-digit number. We take the 8 digits one by one, from the right to the left. We multiply each digit by the power of 2 corresponding to its position. We take these powers and we sum them. We calculate the remainder of this sum by 11. We take this remainder and we calculate its remainder by 10. The result must equal the check digit.

The program will ask the user for TIN and will respond whether it is correct or not. For example:
Enter Tax Identification Number: 090034337
Tax Identification Number valid.

Enter Tax Identification Number: 090034336
Tax Identification Number not valid.